+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 - SC2254: Suggest quoting expansions in case statements
+- SC2255: Suggest using `$((..))` in `[ 2*3 -eq 6 ]`
 
 ## v0.7.0 - 2019-07-28
 ### Added