+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ### Fixed
 - `-f diff` no longer claims that it found more issues when it didn't
 - Known empty variables now correctly trigger SC2086
+- ShellCheck should now be compatible with Cabal 3
 
 ### Added
 - SC2254: Suggest quoting expansions in case statements