+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## v0.7.1 - soon
+## v0.7.1 - 2020-04-04
 ### Fixed
 - `-f diff` no longer claims that it found more issues when it didn't
 - Known empty variables now correctly trigger SC2086
@@ -7,6 +7,7 @@
   called with `builtin`
 
 ### Added
+- SC1136: Warn about unexpected characters after ]/]]
 - SC2254: Suggest quoting expansions in case statements
 - SC2255: Suggest using `$((..))` in `[ 2*3 -eq 6 ]`
 - SC2256: Warn about translated strings that are known variables
@@ -16,6 +17,7 @@
 ### Changed
 - SC2230: 'command -v' suggestion is now off by default (-i deprecate-which)
 - SC1081: Keywords are now correctly parsed case sensitively, with a warning
+
 
 ## v0.7.0 - 2019-07-28
 ### Added